requirements.txt
# asgiref==3.8.1
# atomicwrites==1.4.1
# attrs==25.3.0
# certifi==2025.1.31
# cffi==1.17.1
# charset-normalizer==2.0.12
# colorama==0.4.6
# cryptography==44.0.2
# defusedxml==0.7.1
# Django==2.2.16
# #django-filter==25.1
# #django-filter
# djangorestframework==3.15.2
# djangorestframework_simplejwt==5.5.0
# djoser==2.3.1
# idna==3.10
# iniconfig==2.1.0
# oauthlib==3.2.2
# packaging==24.2
# pillow==11.1.0
# pluggy==0.13.1
# py==1.11.0
# pycparser==2.22
# PyJWT==2.9.0
# pytest==6.2.4
# pytest-django==4.4.0
# pytest-pythonpath==0.7.3
# python3-openid==3.2.0
# pytz==2025.2
# requests==2.26.0
# requests-oauthlib==2.0.0
# setuptools==78.1.0
# social-auth-app-django==5.4.3
# social-auth-core==4.5.6
# sqlparse==0.5.3
# toml==0.10.2
# tzdata==2025.2
# urllib3==1.26.20

Django==3.2.16
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow
PyJWT==2.1.0
requests==2.26.0
django-filter==2.4.0
djoser==2.1.0
setuptools